+++ utils/sendWithdrawEthereumRequest.ts
@@ -1,6 +1,10 @@
-import { BridgedEthereumToken } from "@/types";
-import { Balance, getBridgeContract } from "@/utils";
-import getERC20PegContract from "@/utils/getERC20PegContract";
+import { BridgedEthereumToken, HistoricalEventProof } from "@/types";
+import {
+	Balance,
+	EthereumTransaction,
+	getERC20PegContract,
+	getBridgeContract,
+} from "@/utils";
 import { Api } from "@cennznet/api";
 import { EthEventProof } from "@cennznet/api/derives/ethBridge/types";
 import { BigNumber, ethers } from "ethers";
@@ -8,28 +12,28 @@ import { TransactionResponse } from "@ethersproject/abstract-provider";
 
 export default async function sendWithdrawEthereumRequest(
 	api: Api,
-	eventProof: EthEventProof,
+	eventProof: EthEventProof | HistoricalEventProof,
 	transferAmount: Balance,
 	transferAsset: BridgedEthereumToken,
 	ethereumAddress: string,
-	signer: ethers.Signer,
-	blockHash?: string
-): Promise<TransactionResponse | "cancelled"> {
-	const notaryKeys = !!blockHash
-		? ((
-				await api.query.ethBridge.notaryKeys.at(blockHash)
-		  ).toJSON() as string[])
-		: ((await api.query.ethBridge.notaryKeys()).toJSON() as string[]);
+	signer: ethers.Signer
+): Promise<EthereumTransaction> {
+	let validators: string[];
+	if (!eventProof.validators) {
+		const notaryKeys = (
+			await api.query.ethBridge.notaryKeys()
+		).toJSON() as string[];
 
-	const validators = notaryKeys.map((validator) => {
-		if (
-			validator ===
-			"0x000000000000000000000000000000000000000000000000000000000000000000"
-		)
-			return ethers.constants.AddressZero;
+		validators = notaryKeys.map((validator) => {
+			if (
+				validator ===
+				"0x000000000000000000000000000000000000000000000000000000000000000000"
+			)
+				return ethers.constants.AddressZero;
 
-		return ethers.utils.computeAddress(validator);
-	});
+			return ethers.utils.computeAddress(validator);
+		});
+	}
 
 	const bridgeContract = getBridgeContract<"OnBehalf">(signer);
 	const pegContract = getERC20PegContract<"OnBehalf">(signer);
@@ -38,23 +42,32 @@ export default async function sendWithdrawEthereumRequest(
 		transferAsset.address,
 		transferAmount.toBigNumber(),
 		ethereumAddress,
-		{ ...eventProof, validators },
+		{ ...eventProof, validators: eventProof.validators ?? validators },
 		{ value: verificationFee }
 	);
-
-	try {
-		const tx: TransactionResponse = await pegContract.withdraw(
+	const tx = new EthereumTransaction();
+	pegContract
+		.withdraw(
 			transferAsset.address,
 			transferAmount.toBigNumber(),
 			ethereumAddress,
-			{ ...eventProof, validators },
+			{ ...eventProof, validators: eventProof.validators ?? validators },
 			{ value: verificationFee, gasLimit: (gasFee.toNumber() * 1.02).toFixed() }
-		);
+		)
+		.then((pegTx: TransactionResponse) => {
+			tx.setHash(pegTx.hash);
+			return pegTx.wait(2);
+		})
+		.then(() => {
+			tx.setSuccess();
+		})
+		.catch((error) => {
+			if (error?.code === 4001) {
+				tx.setCancel();
+				return;
+			}
+			tx.setFailure(error?.code);
+		});
 
-		await tx.wait();
-		return tx;
-	} catch (error) {
-		if (error?.code === 4001) return "cancelled";
-		throw error;
-	}
+	return tx;
 }