+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## [1.1.0] - 2022-04-28
+
+- Update UI when transaction is in progress [#156](https://github.com/cennznet/app-hub/pull/156)
+- Update UI/UX for historical withdrawals [#171](https://github.com/cennznet/app-hub/pull/171)
+- Ask user to confirm before exiting if there is a transaction pending [#173](https://github.com/cennznet/app-hub/pull/173)
+
 ## [1.0.2] - 2022-04-01
 
 - Add confirming state to the Bridge [#151](https://github.com/cennznet/app-hub/pull/151)